US - 9 - backend (edit test, try to fix the edit Activities function/methodes) - frontend (fix useActivities axios fail, edit routing)
