+++ frontend/src/App.tsx
@@ -1,47 +1,35 @@
 import "./App.css";
-import HobbyList from "./components/Hobby/HobbyList.tsx";
 import Header from "./components/Header";
-import Button from "@mui/material/Button";
-import {Route, Routes, useNavigate, useParams} from "react-router-dom";
+
+import {Route, Routes, useNavigate} from "react-router-dom";
 import AddForm from "./components/Hobby/AddForm.tsx";
 import useHobbies from "./hooks/useHobbies.ts";
 import styled from "@emotion/styled";
-import HobbyDetail from "./components/Activity/HobbyDetail.tsx";
-import ActivityAddForm from "./components/Activity/ActivityAddForm.tsx";
-import ActivityItem from "./components/Activity/ActivityItem.tsx";
 import CalendarMonthIcon from '@mui/icons-material/CalendarMonth';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import {BottomNavigation, BottomNavigationAction, Paper} from "@mui/material";
 import { useState} from "react";
-import InAppPurchase from "./components/InAppPurchase.tsx";
 import InfoTab from "./components/ProfilePage/InfoTab.tsx";
 import StatisticTab from "./components/ProfilePage/StatisticTab.tsx";
-import Calendar from "./components/Calendar/Calendar.tsx";
 import CalendarActivityAddForm from "./components/Activity/CalendarActivityAddForm.tsx";
 import useActivities from "./hooks/useActivities.ts";
 import LoginForm from "./components/User/LoginForm.tsx";
 
 import RegisterForm from "./components/User/RegisterForm.tsx";
-
 import SettingsTab from "./components/ProfilePage/SettingsTab.tsx";
 import HomePage from "./components/HomePage.tsx";
 import useUser from "./hooks/useUser.ts";
 import ProtectedRoutes from "./components/ProtectedRoutes.tsx";
+import Hobby_Routing from "./Hobby_Routing.tsx";
 export default function App() {
     const navigate = useNavigate();
     const colors = ['coral', 'lightblue', 'cornflowerblue', 'lightgreen', 'seagreen', 'pink', 'mediumpurple', 'orange', 'tomato', 'peachpuff'];
     const [value, setValue] = useState(0);
 
-    const {hobbies, handleAddHobby, handleEditHobbyName, handleEditHobbyColor, handleDeleteHobby, hobby} = useHobbies();
-    const hobbyId= hobby?.hobbyId;
-    const { handleAddActivityToHobby, handleEditActivity, handleDeleteActivity} = useActivities();
+    const {hobbies, hobby, handleAddHobby} = useHobbies();
+    const { handleAddActivityToHobby} = useActivities();
 
-
-    const { activityId} = useParams();
-    const {user, userId, handleLogin, handleRegister, handleLogout} = useUser();
-
-    const selectedHobby = hobbies.find((hobby) => hobby.hobbyId === hobbyId);
-    const selectedActivity = selectedHobby ? selectedHobby.activities.find((activity) => activity.activityId === activityId) : undefined;
+    const { user, userId, handleLogin, handleRegister, handleLogout} = useUser();
 
     const handleProfileIconClick = () => {
         navigate("/profile/info");
@@ -56,63 +44,20 @@ export default function App() {
             <Header/>
             <main>
                 <Routes>
-                    <Route path="/" element={<HomePage />}/>
-                    <Route path="/login" element={<LoginForm onLogin={handleLogin} />} />
+                    <Route path={"/"} element={<HomePage />}/>
+                    <Route path={"/login"} element={<LoginForm onLogin={handleLogin} />} />
                     <Route path="/register" element={<RegisterForm onRegister={handleRegister} />} />
 
-
                     <Route element={<ProtectedRoutes user={user} />}>
-                        <Route path="/hobbies"
-                               element={(
-                                   <>
-                                       <Calendar />
+                        <Route path={"/hobbies"} element={<Hobby_Routing/>} />
 
-                                       <StyledH2>{user}s Hobby List</StyledH2>
-                                       <StyledButtonAdd variant="contained" disableElevation onClick={() => navigate('/add')}>
-                                           +
-                                       </StyledButtonAdd>
+                        <Route path={"profile/*"} element={<InfoTab />} />
+                        <Route path={"profile/info"} element={<InfoTab />} />
+                        <Route path={"profile/stats"} element={<StatisticTab />} />
+                        <Route path={"/profile/settings"} element={<SettingsTab user={user} userId={userId} onLogout={handleLogout}/>} />
 
-                                       <HobbyList hobbies={hobbies} colors={colors} onEditHobbyName={handleEditHobbyName}
-                                                  onEditHobbyColor={handleEditHobbyColor}
-                                                  onDeleteHobby={handleDeleteHobby}/>
-                                   </>
-                               )}
-                        />
-
-                    <Route path="profile/info" element={<InfoTab />} />
-                    <Route path="profile/stats" element={<StatisticTab />} />
-                        <Route path="/profile/settings" element={<SettingsTab user={user} userId={userId} onLogout={handleLogout}/>} />
-
-                        <Route path="profile/*" element={<InfoTab />} />
-
-                    <Route path="/app" element={<InAppPurchase/>}/>
-                    <Route path="/add" element={<AddForm onAddHobby={handleAddHobby} colors={colors}/>}/>
-
-                    <Route path="/calendar/add/" element={<CalendarActivityAddForm onAddActivity={handleAddActivityToHobby} hobbies={hobbies} hobby={hobby}/>} />
-                        <Route
-                        path="/hobby/:hobbyId/activities"
-                        element={<HobbyDetail colors={colors} hobby={hobby}/>}
-                    />
-                    <Route
-                        path="/hobby/:hobbyId/activities/add"
-                        element={<ActivityAddForm onAddActivity={handleAddActivityToHobby} color={selectedHobby?.color ?? colors[0]}/>}
-                    />
-                    <Route
-                        path="/hobby/:hobbyId/activities/:activityId"
-                        element={
-                            selectedHobby && selectedActivity ? (
-                                <ActivityItem
-                                    activity={selectedActivity}
-                                    hobby={selectedHobby}
-                                    colors={colors}
-                                    onEditActivity={handleEditActivity}
-                                    onDeleteActivity={handleDeleteActivity}
-                                />
-                            ) : (
-                                <div>Invalid hobby or activity ID.</div>
-                            )
-                        }/>
-
+                        <Route path={"/add"} element={<AddForm onAddHobby={handleAddHobby} colors={colors}/>}/>
+                        <Route path={"/calendar/add/"} element={<CalendarActivityAddForm onAddActivity={handleAddActivityToHobby} hobbies={hobbies} hobby={hobby}/>} />
                     </Route>
 
                 </Routes>
@@ -135,24 +80,6 @@ export default function App() {
     );
 }
 
-const StyledH2 = styled.h3`
-  margin-top: 0;
-  margin-right: 5rem;
-  padding-top: 1rem;
-  padding-bottom: 0.5rem;
-  border-top: 4px solid black;
-  width: 120%;
-`;
-
-const StyledButtonAdd = styled(Button)`
-  margin: -3rem 0.2rem -6rem 60%;
-  background-color: black;
-  font-size: 25px;
-  &:hover {
-    background-color: darkorange;
-  }
-`;
-
 const StyledPaper = styled(Paper)`
   position: fixed;
   bottom: 0;
@@ -170,14 +97,14 @@ const StyledBottomNavigation = styled(BottomNavigation)`
   justify-content: space-around;
   align-items: center;
   background-color: black;
-  height: 2rem; 
-  position: fixed; 
-  bottom: 0; 
-  left: 0; 
+  height: 2rem;
+  position: fixed;
+  bottom: 0;
+  left: 0;
   right: 0;
   margin: 0;
 `;
 
 const StyledBottomNavigationAction = styled(BottomNavigationAction)`
- color: white;
+  color: white;
 `;
